Forcing C11 and C++14 standards and updating the main library
Using the latest version, at this commit date, of MyyOpenGLHelpers.
Also forcing the C standard to C11 and the C++ standard to C++14 in
order to use all the new functionnalities of these languages without
issue.

Now, currently the main issue with not defining the C standard, was
that GCC was spouting errors about variables definitions in 'for'
loops.

This could be solved by bumping the C standard requirement to C99, but
quite frankly, any modern C compiler supports C11, so let be it.

Signed-off-by: Myy <myy@miouyouyou.fr>