Added the first Ruby tool
This *might* be replaced by a C program, in order to avoid adding more
dependencies.

The tool must be executed from the project root directory like this :

    ruby tools/ShadersPacker.rb

And will generate :
- A shaders.pack in data/
- A C helper header in src/generated/opengl/shaders_infos.h

This header will provide the following :
- Enumerators for shaders files using the following format :
 - glsl_file_#{base_filename}_vsh
 - glsl_file_#{base_filename}_fsh
- Enumerators for (guessed) program names using the following format :
 - glsl_program_#{program_name}
 Note that program_name is guessed using the filenames. For each
 different filename, a program_name will be generated.
 The convention is that for each program, a .vsh and .fsh file with the
 same name must be present in the shaders/ folder.
 If you wish to the same shaders for multiple programs, create symlinks.
- Enumerators for each shader's uniforms, using the following format :
 - #{program_name}_shader_unif_#{uniform_name}
For each program name generated, an attribute enum will be generated,
with members using the following format :
 - #{program_name}_shader_attr_#{attribute_name}

Along with data structures definitions currently usable with :
- glhShadersPackLoader
- glhBuildAndSaveSimpleProgram
And specifically fitted to load the generated pack.

The shaders.pack might be renamed in the future, as it only contains
shaders metadata. shaders.metadata maybe ?

Signed-off-by: Myy <myy@miouyouyou.fr>