Replacing the old Ruby packer draft with a new ... draft
The code is still very ugly since I have almost no time and
in a big hurry.

However, the shaders data are packed correctly and there's now
ONE FILE FOR ALL when it comes to shaders. This should make a
few things simpler.

A lot of things needs to be factorized, renamed, ...
But for now, I'll just get some sleep.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>