So many things at once...
The font packer is now redone for... almost nothing.
I just want to avoid relying too much on Ruby since I'd like to use
these tools on many platforms, and some platforms are not very Ruby
friendly.
Also, relying on gems requiring binary libraries doesn't help.

However, that means that an old ~400 lines Ruby program turned into
a 1200 lines C monster.
Now the new font packer handle multiple fonts and load the data from
a text file but... that's not much.

The ShadersPacker.rb has still rewritten in Ruby. It now generates
a pack file, from programs defined in a .ini file. The programs and
all their uniforms/attribute names are packed into a single file
that can be loaded at once, or used to load a specific program by
name.
The script generates a header file that defines enumerators to
reference the bound attributes, and a data structure that receives
all the programs and uniforms ID during the loading phase, in C.

The vectors helpers are now being integrated with the main helpers.

And myy.c now displays the weirdest and stupidest line ever on
screen, to show the font display working.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>