Scrolling menus limited to a specific screen portion !
Thanks to the Stencil mechanisms, it's quite easy to limit drawing to
a specific area of the screen. Using this system, it's quite easy to
display large scrollable menus into a smaller screen area.

A new font was added, in order to make things a little nicer to the
eye.

Now, I'll have to add the functions menus.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>