Hitboxes are so nice ! So let's add a text editing module instead.
That comes from a simple idea :
* Let's forget about text handling, we'll just focus on Android first.

Then I started to test the Android Virtual Keyboard support,
stumbled onto the VERY WEIRD license on the Android SDK that states
a lot of things that completely contradict the Apache license.

And realized that I might have to first implement some "start"
of a text edit thing for X11.
Which led me to X11 international input handling... Oh boy, what
a lovely ride, into the madness of the X11 protocol and it's current
implementation with XCB, and the good old Xlib.
Took me a few hours to understand how the XIM part of the Xlib were
able to "convert text inserted" without storing it somewhere.
Spoiler alert : it stores it in the "private data" section, located
at the "IM" address. I wrote a lengthy documentation about it AND
... I lost it by closing the editor too abrubtly, and ignoring the
various "There are unsaved changes" warnings... Ugh...

Anyway, after refocusing back on this project, I was able to add
some "text editing module" prototype and... IT WORKS !!! YAY !!!
It's still in a "0.alpha1" state but, it's kind of useable.

So, now I'll still try to do the Android part, release a 0alpha1
of the whole graphical assembler, and then focus on a making a
nice XML documentation editor, in order to start writing the
documentation of the various parts used in this program, because
I'm starting to be overloaded by the little parts started here
and there, that have zero documentation...

So... YEAH ! THE BEGINNING OF TEXT EDITING ! ... In 2019...

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>