The android input system is working !
Now, some things needs to be clearly rewritten before advancing.
Seriously, the widgets should have simple and clean functions to
interact with them. Any GPU handling to interact with a widget
is beyond retarded.

Also, we clearly need to pass some global state to the every
myy_ function. This should include, the current frame, the
last time delta, the width and height of the current scene
and a pointer to some platform specific state and another
to user-specific state.

Storing these states here and there, in a random fashion make
things more difficult that they should be.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>