Cleaned up the main file by putting widgets definitions away
Into src/widgets ...

I already done that once and there are issues with splitting things
too much.
Still, the myy.c went from 1K+ lines to ~300 lines... making it a
little more clearer.

Now, all widgets are drawn from the bottom left, making things easier
when trying to align them on the screen.
The reason for this alignment is that text is drawn from the bottom
left. So everything has to follow.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>